+++ packages/plugin-core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @granite-js/plugin-core
 
+## 0.1.22
+
+### Patch Changes
+
+- @granite-js/utils@0.1.22
+
+## 0.1.21
+
+### Patch Changes
+
+- @granite-js/utils@0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- @granite-js/utils@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- e9a3daf: missing Metro configuration compatibility
+  - @granite-js/utils@0.1.19
+
+## 0.1.18
+
+### Patch Changes
+
+- @granite-js/utils@0.1.18
+
 ## 0.1.17
 
 ### Patch Changes